feat: move role related methods to role service
