+++ src/main/java/com/skapp/community/peopleplanner/service/impl/PeopleServiceImpl.java
@@ -19,7 +19,6 @@ import com.skapp.community.common.service.EncryptionDecryptionService;
 import com.skapp.community.common.service.UserService;
 import com.skapp.community.common.service.impl.AsyncEmailServiceImpl;
 import com.skapp.community.common.type.LoginMethod;
-import com.skapp.community.common.type.ModuleType;
 import com.skapp.community.common.type.NotificationSettingsType;
 import com.skapp.community.common.type.Role;
 import com.skapp.community.common.util.CommonModuleUtils;
@@ -63,7 +62,6 @@ import com.skapp.community.peopleplanner.payload.request.JobTitleDto;
 import com.skapp.community.peopleplanner.payload.request.NotificationSettingsPatchRequestDto;
 import com.skapp.community.peopleplanner.payload.request.PermissionFilterDto;
 import com.skapp.community.peopleplanner.payload.request.ProbationPeriodDto;
-import com.skapp.community.peopleplanner.payload.request.RoleRequestDto;
 import com.skapp.community.peopleplanner.payload.response.AnalyticsSearchResponseDto;
 import com.skapp.community.peopleplanner.payload.response.EmployeeBulkErrorResponseDto;
 import com.skapp.community.peopleplanner.payload.response.EmployeeBulkResponseDto;
@@ -81,7 +79,6 @@ import com.skapp.community.peopleplanner.payload.response.EmployeeResponseDto;
 import com.skapp.community.peopleplanner.payload.response.EmployeeTeamDto;
 import com.skapp.community.peopleplanner.payload.response.ManagerEmployeeDto;
 import com.skapp.community.peopleplanner.payload.response.ManagingEmployeesResponseDto;
-import com.skapp.community.peopleplanner.payload.response.ModuleRoleRestrictionResponseDto;
 import com.skapp.community.peopleplanner.payload.response.PrimarySecondaryOrTeamSupervisorResponseDto;
 import com.skapp.community.peopleplanner.payload.response.SummarizedEmployeeDtoForEmployees;
 import com.skapp.community.peopleplanner.payload.response.SummarizedManagerEmployeeDto;
@@ -253,7 +250,7 @@ public class PeopleServiceImpl implements PeopleService {
 		}
 
 		// Validate the roles
-		validateRoles(employeeDetailsDto.getUserRoles());
+		rolesService.validateRoles(employeeDetailsDto.getUserRoles());
 
 		// Validate the employee details
 		Validations.validateEmployeeDetails(employeeDetailsDto);
@@ -349,7 +346,7 @@ public class PeopleServiceImpl implements PeopleService {
 
 		Validations.validateQuickAddEmployeeDetails(employeeQuickAddDto);
 
-		validateRoles(employeeQuickAddDto.getUserRoles());
+		rolesService.validateRoles(employeeQuickAddDto.getUserRoles());
 
 		Employee finalEmployee = peopleMapper.employeeQuickAddDtoToEmployee(employeeQuickAddDto);
 		finalEmployee.setAccountStatus(AccountStatus.PENDING);
@@ -423,7 +420,7 @@ public class PeopleServiceImpl implements PeopleService {
 			}
 		}
 
-		validateRoles(employeeUpdateDto.getUserRoles());
+		rolesService.validateRoles(employeeUpdateDto.getUserRoles());
 
 		String employeePreviousName = optionalEmployee.get().getFirstName();
 		String employeePreviousLastName = optionalEmployee.get().getLastName();
@@ -3176,42 +3173,6 @@ public class PeopleServiceImpl implements PeopleService {
 			throw new EntityNotFoundException(PeopleMessageConstant.PEOPLE_ERROR_TEAM_NOT_FOUND);
 		}
 		return employeeTeams;
-	}
-
-	private void validateRoles(RoleRequestDto userRoles) {
-		User currentUser = userService.getCurrentUser();
-
-		if (hasOnlyAdminPermissions(currentUser) && ((userRoles.getAttendanceRole() != null
-				&& validateRestrictedRoleAssignment(userRoles.getAttendanceRole(), ModuleType.ATTENDANCE))
-				|| (userRoles.getPeopleRole() != null
-						&& validateRestrictedRoleAssignment(userRoles.getPeopleRole(), ModuleType.PEOPLE))
-				|| (userRoles.getLeaveRole() != null
-						&& validateRestrictedRoleAssignment(userRoles.getLeaveRole(), ModuleType.LEAVE)))) {
-			throw new ModuleException(PeopleMessageConstant.PEOPLE_ERROR_SUPER_ADMIN_RESTRICTED_ASSIGNING_ROLE_ACCESS);
-		}
-
-		if (Boolean.TRUE.equals(userRoles.getIsSuperAdmin())
-				&& (userRoles.getPeopleRole() != Role.PEOPLE_ADMIN || userRoles.getLeaveRole() != Role.LEAVE_ADMIN
-						|| userRoles.getAttendanceRole() != Role.ATTENDANCE_ADMIN)) {
-			throw new ModuleException(PeopleMessageConstant.PEOPLE_ERROR_SHOULD_ASSIGN_PROPER_PERMISSIONS);
-		}
-	}
-
-	private boolean hasOnlyAdminPermissions(User currentUser) {
-		return Boolean.FALSE.equals(currentUser.getEmployee().getEmployeeRole().getIsSuperAdmin())
-				&& currentUser.getEmployee().getEmployeeRole().getPeopleRole() == Role.PEOPLE_ADMIN;
-	}
-
-	private Boolean validateRestrictedRoleAssignment(Role role, ModuleType moduleType) {
-		ModuleRoleRestrictionResponseDto restrictedRole = rolesService.getRestrictedRoleByModule(moduleType);
-
-		return switch (role) {
-			case PEOPLE_ADMIN, ATTENDANCE_ADMIN, LEAVE_ADMIN -> Boolean.TRUE.equals(restrictedRole.getIsAdmin());
-			case PEOPLE_MANAGER, ATTENDANCE_MANAGER, LEAVE_MANAGER ->
-				Boolean.TRUE.equals(restrictedRole.getIsManager());
-			default -> false;
-		};
-
 	}
 
 }
